pyproject.toml
[project]
name = "zthob"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.15.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.27.0",
    "pillow>=11.3.0",
]
